% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_wrapper.R
\name{analysis_wrapper}
\alias{analysis_wrapper}
\title{Create a function that replicates an analysis for all time series in a
dataset}
\usage{
analysis_wrapper(fun)
}
\arguments{
\item{fun}{the analysis function}
}
\value{
a function that takes in a \code{dataset} and optional arguments, and
returns a data.frame or tibble with the combined results, and an "id"
column with the name of the species
}
\description{
This wrapper takes as input, some analysis function, \code{fun}, and
returns a function that will run that analysis function on all the time
series in a dataset. Some post-processing attempts to handle merging of
the output in a sensible way.
}
\examples{
\dontrun{
  sgs_data <- MATSS::get_sgs_data()
  summarize_dataset <- analysis_wrapper(ts_summary)
  summarize_dataset(sgs_data)
}

}
