% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-retriever-data.R
\name{append_data_citations}
\alias{append_data_citations}
\title{Generate a vector of citations.}
\usage{
append_data_citations(citations = NULL, citation_files)
}
\arguments{
\item{citations}{a vector of strings containing existing citations to append}

\item{citation_files}{a vector of filepaths to the citation files}
}
\value{
a vector of strings containing the citations
}
\description{
Given an existing vector of citations (or the NULL default),
add the citations that are specified in the paths of \code{citation_files}
}
