% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-retriever-data.R
\name{append_retriever_citation}
\alias{append_retriever_citation}
\title{Append citation info to a formatted dataset}
\usage{
append_retriever_citation(formatted_data, path)
}
\arguments{
\item{formatted_data}{a dataset that already follows the `MATSS`` standard}

\item{path}{where to load the raw data files from}
}
\value{
the same dataset, with the citation appended to \code{metadata}
}
\description{
Given an existing formatted dataset, and the path to the
downloaded dataset, from retriever, and via \code{import_retriever_data()},
read in the citation info and add it to the metadata for the dataset
}
