% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plans.R
\name{build_bbs_datasets_plan}
\alias{build_bbs_datasets_plan}
\title{Generate a Drake Plan for BBS Datasets}
\usage{
build_bbs_datasets_plan(bbs_ts_data)
}
\arguments{
\item{data_path}{}
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands 
  for gathering BBS datasets
}
\description{
Generate a Drake Plan for BBS Datasets
}
