% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plans.R
\name{build_biotime_datasets_plan}
\alias{build_biotime_datasets_plan}
\title{Generate a Drake Plan for Biotime Datasets}
\usage{
build_biotime_datasets_plan(path = get_default_data_path(),
  data_subset = NULL)
}
\arguments{
\item{path}{where to get the downloaded retriever datasets}
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands
for gathering Biotime datasets
}
\description{
Generate a Drake Plan for Biotime Datasets
}
