% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plans.R
\name{build_datasets_plan}
\alias{build_datasets_plan}
\title{Generate a Drake Plan for Datasets}
\usage{
build_datasets_plan()
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands 
  for gathering datasets
}
\description{
Generate a Drake Plan for Datasets
}
