% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_data.R
\name{build_gpdd_datasets_plan}
\alias{build_gpdd_datasets_plan}
\title{Generate a Drake Plan for GPDD Datasets}
\usage{
build_gpdd_datasets_plan(path = get_deault_data_path(), data_subset = NULL)
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands
for gathering GPDD datasets
}
\description{
Generate a Drake Plan for GPDD Datasets
}
