% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_references.R
\name{build_references_plan}
\alias{build_references_plan}
\title{Generate a Drake Plan for dataset references}
\usage{
build_references_plan(datasets, ...)
}
\arguments{
\item{datasets}{a drake plan listing the datasets to be analyzed}

\item{...}{arguments to be passed to \code{drake::\link[drake]{drake_plan}}}
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands
for all the references and the combined vector
}
\description{
Given N datasets, extract the citation from each one, and then
combine them into a single vector and remove duplicates.
}
