% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_analyses.R, R/plan_data.R
\name{build_retriever_datasets_plan}
\alias{build_retriever_datasets_plan}
\title{Generate a Drake Plan for retriever datasets}
\usage{
build_retriever_datasets_plan(path = get_default_data_path())

build_retriever_datasets_plan(path = get_default_data_path())
}
\arguments{
\item{path}{where to get the downloaded retriever datasets}
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands
for retriever downloaded datasets

a drake plan (i.e. a tibble) specifying the targets and commands
for retriever downloaded datasets
}
\description{
Generate a Drake Plan for retriever datasets

Generate a Drake Plan for retriever datasets
}
