% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-retriever-data.R
\name{check_default_data_path}
\alias{check_default_data_path}
\title{Check if a default data path is set}
\usage{
check_default_data_path()
}
\description{
See \code{portalr::\link[portalr]{check_default_data_path}} for details.
}
