% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_bbs.R
\name{combine_bbs_subspecies}
\alias{combine_bbs_subspecies}
\title{Combine subspecies into their common species}
\usage{
combine_bbs_subspecies(bbs_data_table, species_table)
}
\arguments{
\item{bbs_data_table}{main bbs data table}

\item{species_table}{table of species for BBS}
}
\description{
Modified from \url{https://github.com/weecology/bbs-forecasting}
}
