% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biotime_data.R
\name{correct_biotime_dataset}
\alias{correct_biotime_dataset}
\title{Correct and clean specific datasets}
\usage{
correct_biotime_dataset(raw_data, dataset_id = 10)
}
\arguments{
\item{raw_data}{The raw data for a specific dataset_id}

\item{dataset_id}{the study_id}
}
\value{
a corrected version of \code{raw_data}
}
\description{
Correct and clean specific datasets
}
\details{
For \code{dataset_id = 54}, it appears that day and month were sometimes
interchanged. Since there did not seem to be measurements after August in
any given year otherwise, we use that to filter and swap \code{day} and \code{month}.
}
