% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_MATSS_compendium.R
\name{create_MATSS_compendium}
\alias{create_MATSS_compendium}
\title{Create a research compendium for MATSS}
\usage{
create_MATSS_compendium(path, fields = NULL,
  rstudio = rstudioapi::isAvailable(), open = interactive())
}
\value{
Path to the newly created package, invisibly.
}
\description{
Create an R package, using \code{\link[usethis]{use_package}},
then perform the following additional actions:
\itemize{
\item add various package dependencies, including "MATSS"
\item add template analysis and pipeline files
}
}
