% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATSS-package.R
\docType{data}
\name{dragons}
\alias{dragons}
\title{Example dataset}
\format{A list with 3 elements:
\describe{
\item{abundance}{a data.frame with abundances for 3 dragons}
\item{covariates}{a data.frame with times of observations and effort and precip data}
\item{metadata}{a list with:
\code{timename} - the name of the time column in covariates,
\code{period} - the gap between successive observations,
\code{authors} - the authors of the dataset,
\code{species_table} - information about the species observed}
}}
\usage{
dragons
}
\description{
A dataset containing example timeseries for some dragons.
}
\keyword{datasets}
