% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_cleaning_functions.R
\name{filter_bbs_species}
\alias{filter_bbs_species}
\title{Filter poorly sampled BBS species}
\usage{
filter_bbs_species(bbs_data_table, species_table)
}
\arguments{
\item{bbs_data_table}{main bbs data table}

\item{species_table}{table of species for BBS}
}
\value{
dataframe, filtered version of initial dataframe
}
\description{
Modified from https://github.com/weecology/bbs-forecasting/blob/master/R/forecast-bbs-core.R
}
\details{
Removes waterbirds, shorebirds, owls, kingfishers, knightjars,
dippers. These species are poorly sampled due to their aquatic or
noctural nature. Also removes taxa that were either partially unidentified
(e.g. "sp.") or were considered hybrids (e.g. "A x B") or were listed as more
than one species (e.g. "A / B")
}
