% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{filter_ts}
\alias{filter_ts}
\title{Filter BBS to specified time series period and number of samples}
\usage{
filter_ts(bbs_data, start_yr, end_yr, min_num_yrs)
}
\arguments{
\item{bbs_data}{dataframe that contains BBS site_id and year columns}

\item{start_yr}{num first year of time-series}

\item{end_yr}{num last year of time-series}

\item{min_num_yrs}{num minimum number of years of data between start_yr & end_yr}
}
\value{
dataframe with original data and associated environmental data
}
\description{
Modified from https://github.com/weecology/bbs-forecasting 
and https://github.com/weecology/MATSS-community-change
}
