% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plans.R
\name{generate_analyses_section}
\alias{generate_analyses_section}
\title{generate_analyses_section}
\usage{
generate_analyses_section()
}
\value{
the analyses chunk
}
\description{
generate_analyses_section
}
