% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plans.R
\name{generate_methods}
\alias{generate_methods}
\title{generate_methods}
\usage{
generate_methods()
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands 
  for analysis methods
}
\description{
generate_methods
}
