% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biotime_data.R
\name{get_biotime_data}
\alias{get_biotime_data}
\title{Get Biotime time-series data}
\usage{
get_biotime_data(dataset, path = get_default_data_path())
}
\arguments{
\item{dataset}{Dataset number to retrieve}

\item{path}{Data storage path}
}
\value{
list of two dataframes (one with abundance data, the other with covariate data)
and one list of metadata.
}
\description{
Original data found here http://biotime.st-andrews.ac.uk/home.php
}
\examples{
\dontrun{
  get_biotime_data(dataset = 321)
}
}
