% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_biotime.R
\name{get_biotime_dataset_ids}
\alias{get_biotime_dataset_ids}
\alias{prepare_biotime_data}
\title{Return BioTime dataset ids for individual loading}
\usage{
get_biotime_dataset_ids(path = get_default_data_path(),
  data_subset = NULL, do_processing = FALSE,
  force_reprocessing = FALSE)

prepare_biotime_data(path = get_default_data_path(),
  data_subset = NULL)
}
\arguments{
\item{path}{where to get the downloaded retriever datasets}

\item{data_subset}{optional, a subset of the Biotime study_ids to use
(to speed up development). As c(1:X)}

\item{do_processing}{whether to process the datasets if necessary}
}
\value{
vector of dataset ids in the processed set of files

vector of dataset ids in the processed set of files
}
\description{
Retrieve the dataset ids from processed BioTime files. If the
processed files do not exist, and \code{do_processing == TRUE}, then we also
load the raw BioTime database and process the necessary datasets, too.

\code{prepare_biotime_data} is a thin wrapper around
\code{get_biotime_dataset_ids()} for processing BioTime dataset
}
\examples{
\dontrun{
  get_biotime_dataset_ids()
}
\dontrun{
  prepare_biotime_data()
}
}
