% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cowley_data.R
\name{get_cowley_lizards}
\alias{get_cowley_lizards}
\title{Read in the cowley lizard community data from a txt file

Import cowley lizard data from data files}
\usage{
get_cowley_lizards()
}
\value{
list of two dataframes (one with abundance data, the other with 
  covariate data) and one list of metadata.
}
\description{
Read in the cowley lizard community data from a txt file

Import cowley lizard data from data files
}
