% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_retriever_data.R
\name{get_default_data_path}
\alias{get_default_data_path}
\title{What is the default data path?}
\usage{
get_default_data_path(fallback = "~")
}
\arguments{
\item{fallback}{the default value to use if the setting is missing}
}
\description{
See \code{portalr::\link[portalr]{get_default_data_path}} for details.
}
