% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-checks.R
\name{get_effort_from_data}
\alias{get_effort_from_data}
\title{extract the effort from a formatted data structure}
\usage{
get_effort_from_data(data)
}
\arguments{
\item{data}{a formatted data structure}
}
\value{
a \code{numeric} vector containing the effort
}
\description{
extract the effort from a formatted data structure
}
