% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gpdd_data.R
\name{get_gpdd_data}
\alias{get_gpdd_data}
\title{Create GPDD population time-series data}
\usage{
get_gpdd_data(min_num_yrs = 10, location_id, timeperiod_id)
}
\arguments{
\item{min_num_yrs}{minimum number of years of data}

\item{location_id}{Location code of data to return}

\item{timeperiod_id}{Sampling timescale code of data to return
(some datasets provide at more than one scale)}
}
\value{
list of two dataframes (one with abundance data, the other with covariate data)
and one list of metadata.
}
\description{
Selects sites containing at
least min_num_yrs of data samples during that period.
}
\examples{
\dontrun{
  get_gpdd_data(location_id=83, timeperiod_id=408)
}
}
