% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jornada_data.R
\name{get_jornada_data}
\alias{get_jornada_data}
\title{get Jornada rodent data

Import Jornada rodent abundance from data files}
\usage{
get_jornada_data()
}
\value{
list of two dataframes (one with abundance data, the other with covariate data) 
  and one list of metadata.
}
\description{
get Jornada rodent data

Import Jornada rodent abundance from data files
}
