% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sanparks_data.R
\name{get_karoo_data}
\alias{get_karoo_data}
\title{get Karoo ungulate data

Import and clean Karoo abundance from data files}
\usage{
get_karoo_data()
}
\value{
list of two dataframes (one with abundance data, the other with
  covariate data), and one list of metadata.
}
\description{
get Karoo ungulate data

Import and clean Karoo abundance from data files
}
