% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sanparks_data.R
\name{get_kruger_data}
\alias{get_kruger_data}
\title{get Kruger National Park ungulate data

Import and clean Kruger National Park abundance from data files}
\usage{
get_kruger_data()
}
\value{
list of two dataframes (one with abundance data, the other with
covariate data), and one list of metadata.
}
\description{
get Kruger National Park ungulate data

Import and clean Kruger National Park abundance from data files
}
