% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_mtquad.R
\name{get_mtquad_data}
\alias{get_mtquad_data}
\title{get Montana plant quad time-series data}
\usage{
get_mtquad_data(
  path = file.path(get_default_data_path(), "mapped-plant-quads-mt")
)
}
\arguments{
\item{path}{where to load the raw data files from}
}
\value{
list of abundance, covariates, and metadata
}
\description{
get Montana plant quad time-series data
}
