% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-function_portal.R
\name{get_portal_rodents}
\alias{get_portal_rodents}
\title{get portal rodent data

Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod 
tempor incididunt ut labore et dolore magna aliqua.}
\usage{
get_portal_rodents(time_or_plots = "plots", treatment = "control",
  type = "Rodents")
}
\arguments{
\item{time_or_plots}{select whether to: (1) "time" == get the data for the 
entire timespan of the experiment, or (2) "plots" == just the time period 
with consistent treatments}

\item{treatment}{"control" or "exclosure" treatments}

\item{type}{type of animals to get}
}
