% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retriever_data.R
\name{get_retriever_data}
\alias{get_retriever_data}
\title{Fetch a dataset from data retriever}
\usage{
get_retriever_data(name, save = FALSE)
}
\arguments{
\item{name}{Name of dataset to load}

\item{save}{Logical If TRUE, save to data directory, in addition to loading}
}
\value{
A list, all data tables associated with named dataset
}
\description{
Load data, and optionally save to data directory
}
\examples{
get_retriever_data(portal, save = FALSE)

}
