% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_shortgrass_steppe.R
\name{get_sgs_data}
\alias{get_sgs_data}
\title{get Shortgrass Steppe rodent data}
\usage{
get_sgs_data(
  path = file.path(get_default_data_path(), "shortgrass-steppe-lter")
)
}
\arguments{
\item{path}{where to load the raw data files from}
}
\value{
list of abundance, covariates, and metadata
}
\description{
get Shortgrass Steppe rodent data
}
