% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_format.R
\name{get_times_from_data}
\alias{get_times_from_data}
\title{extract the times from a formatted data structure}
\usage{
get_times_from_data(data)
}
\arguments{
\item{data}{a formatted data structure}
}
\value{
a \code{numeric} or \code{Date} vector containing the times
}
\description{
extract the times from a formatted data structure
}
