% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-processing.R
\name{has_integer_times}
\alias{has_integer_times}
\title{Check if a dataset has integer times}
\usage{
has_integer_times(data)
}
\arguments{
\item{data}{dataset to check}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if a dataset has integer times
}
\details{
If the times are already integer or Date, true. Otherwise FALSE,
with a message if times are missing, or if times could potentially be
rounded.
}
