% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-processing.R
\name{has_missing_samples}
\alias{has_missing_samples}
\alias{is_fully_sampled}
\title{Check for missing samples}
\usage{
has_missing_samples(data, period = NULL, tol = 1e-06, check_covariates = FALSE)
}
\arguments{
\item{data}{dataset to check}

\item{period}{period to check the times against (if \code{NULL}, first check to
see if there is a known \code{period} set in the metadata, otherwise assumes 1)}

\item{tol}{tolerance for the period}

\item{check_covariates}{\code{TRUE} or \code{FALSE} (whether to check covariates, too)}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Some analyses may require evenly sampled data without missing
values. \code{has_missing_samples} checks that the dataset is equitimed, and
then for missing values within \code{abundance} (and optionally, \code{covariates})

\code{is_full_sampled()} does the same check, but returns \code{TRUE} if there are
NO missing samples.
}
