% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_retriever_data.R
\name{import_retriever_data}
\alias{import_retriever_data}
\title{Import a dataset downloaded from data retriever}
\usage{
import_retriever_data(dataset, data_path = "data")
}
\arguments{
\item{dataset}{the name of the dataset that you wish to download}

\item{data_path}{the overarching folder in which to import datasets from}
}
\description{
a function to import a previously downloaded dataset from 
  rdataretriever
}
\examples{
\dontrun{
  import_retriever_data("veg-plots-sdl")
}
}
