% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-retriever-data.R
\name{install_retriever_data}
\alias{install_retriever_data}
\alias{import_retriever_data}
\title{Download a dataset from data retriever}
\usage{
install_retriever_data(dataset, path = get_default_data_path(),
  force_install = FALSE)

import_retriever_data(dataset, path = get_default_data_path())
}
\arguments{
\item{dataset}{the name of the dataset that you wish to download}

\item{path}{the overarching folder in which to download datasets}

\item{force_install}{whether to install the dataset if the correctly named
folder already exists}
}
\description{
\code{install_retriever_data} downloads retriever datasets and
is a wrapper around rdataretriever::install

\code{import_retriever_data} loads a previously downloaded
retriever dataset
}
\examples{
\dontrun{
  install_retriever_data("veg-plots-sdl")
}
\dontrun{
  import_retriever_data("veg-plots-sdl")
}
}
