% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-processing.R
\name{interpolate_missing_samples}
\alias{interpolate_missing_samples}
\title{Impute missing samples using linear interpolation}
\usage{
interpolate_missing_samples(
  data,
  period = NULL,
  tol = 1e-06,
  interpolate_covariates = FALSE
)
}
\arguments{
\item{data}{dataset to modify}

\item{period}{period to check the times against (if \code{NULL}, first check to
see if there is a known \code{period} set in the metadata, otherwise assumes 1)}

\item{tol}{tolerance for the period}

\item{interpolate_covariates}{\code{TRUE} or \code{FALSE} (whether to do covariates, too)}
}
\value{
the dataset, with interpolated samples
}
\description{
Impute missing samples using linear interpolation
}
\details{
First, check if the data are evenly sampled in time. If not, we
exit early. Next, apply \code{forecast::na.interp()} to each variable that has
non-finite values.
}
