% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{interpolate_obs}
\alias{interpolate_obs}
\title{Interpolate observations}
\usage{
interpolate_obs(obs, times, interp_method = forecast::na.interp, ...)
}
\arguments{
\item{obs}{a vector, matrix, or data.frame of \code{numeric} observations
(within columns) across times (within rows).}

\item{times}{\code{numeric} or \code{Date} vector of timestamps of the
observations.}

\item{interp_method}{\code{character} a function used to interpolate
\code{obs}. Defaults to \code{\link[forecast]{na.interp}}.}

\item{...}{further arguments to be passed to the interpolation method}
}
\value{
Interpolated observation vector.
}
\description{
Interpolate observations based on their timestamps and a
method.
}
