% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_wrapper.R
\name{invoke}
\alias{invoke}
\title{Invoke an analysis on a dataset}
\usage{
invoke(fun, data, ...)
}
\arguments{
\item{fun}{the analysis function}

\item{data}{the dataset}

\item{...}{additional arguments to pass to \code{fun}}
}
\value{
a tibble with these columns:
\tabular{ll}{
\code{results} \tab the output of \code{fun(data)}\cr
\code{metadata} \tab the metadata component of the original dataset\cr
\code{dataset} \tab the name of the dataset\cr
\code{method} \tab the name of the analysis function\cr
\code{args} \tab a list of optional args to \code{method}\cr
}
}
\description{
This function is a helper that, at its core, simply applies the
function to the dataset. The key added functionality is to preserve the
names of the function and the dataset, as well as metadata and any
additional arguments; returning the result in a tibble that is consistent
in format, regardless of what function is actually being invoked.
}
\examples{
\dontrun{
  sgs_data <- MATSS::get_sgs_data()
  invoke(ts_summary, sgs_data)
}

}
