% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-processing.R
\name{is_equitimed}
\alias{is_equitimed}
\alias{is_evenly_sampled}
\title{Check that the times of a dataset are evenly sampled}
\usage{
is_equitimed(data, period = NULL, tol = 1e-06, return_if_time_missing = TRUE)
}
\arguments{
\item{data}{dataset to check}

\item{period}{period to check the times against (if \code{NULL}, first check to
see if there is a known \code{period} set in the metadata, otherwise assumes 1)}

\item{tol}{tolerance for the period}

\item{return_if_time_missing}{default value to return if times are missing}
}
\value{
TRUE or FALSE, invisibly
}
\description{
Check that the times of a dataset are evenly sampled
}
