% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-processing.R
\name{make_equitimed}
\alias{make_equitimed}
\alias{make_evenly_sampled}
\title{Insert rows if necessary so that time series are evenly sampled}
\usage{
make_equitimed(
  data,
  period = NULL,
  tol = 1e-06,
  method = c("mean", "method", "closest"),
  na.rm = TRUE
)
}
\arguments{
\item{data}{dataset to modify}

\item{period}{period to check the times against (if \code{NULL}, first check to
see if there is a known \code{period} set in the metadata, otherwise assumes 1)}

\item{tol}{tolerance for the period}

\item{method}{one of \code{c("mean", "method", "closest")} that determines how
the rows of the original data will get coerced into the output here.}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\value{
the dataset, with rows coerced according to the equitimed time
indices, and additional empty rows inserted if needed
}
\description{
Insert rows if necessary so that time series are evenly sampled
}
\details{
First, \code{get_full_times()} computes the sequence of time index values
at a regular sampling interval of period. These will be the final time
index values for the output. \emph{Some} set of rows of the original dataset
will map to each of these time indices.

The \code{method} argument determines how these rows get coerced:
\describe{
\item{mean}{the values in the rows are averaged together using \code{mean}}
\item{median}{the values in the rows are averaged together using \code{median}}
\item{closest}{the values in the row that is closest in time to the
desired time index are used.}
}
}
