% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-processing.R
\name{make_integer_times}
\alias{make_integer_times}
\title{Add a time variable with integer values for evenly sampled data}
\usage{
make_integer_times(data)
}
\arguments{
\item{data}{dataset to modify}
}
\value{
the dataset, with integer times
}
\description{
Add a time variable with integer values for evenly sampled data
}
\details{
If the times are already integer or Date, we don't do anything. If
the times are numeric, but roundable to integer, we round. Otherwise, we
add a new variable to \code{covariates} from 1:n and designate this variable as
the \code{timename}.
}
