% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATSS-package.R
\docType{package}
\name{MATSS}
\alias{MATSS}
\alias{MATSS-package}
\title{Macroecological Analayses of Time Series Structure}
\description{
Support for macroecological analyses of time series. The intent
of the package is to enable end users to run analyses on a collection of
population and community time series data. Functions are provided to
download and import datasets, produce reproducible workflows using the
\code{drake} package, and for generating research compendia with code and
reports.
}
\keyword{package}
