% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATSS-pipeline.R
\docType{package}
\name{MATSSpipeline}
\alias{MATSSpipeline}
\alias{MATSSpipeline-package}
\title{Macroecological Analayses of Time Series Structure pipeline}
\description{
Analyze the structure of ecological time series.
}
\keyword{package}
