% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{normalize_obs}
\alias{normalize_obs}
\title{Normalize obs to effort}
\usage{
normalize_obs(obs, effort = NULL, obs_per_effort = !is.null(effort))
}
\arguments{
\item{obs}{obseved data}

\item{effort}{effort data}

\item{obs_per_effort}{whether to scale observation by effort}
}
\description{
Normalize obs to effort
}
