% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_datasets.R
\name{plan_datasets}
\alias{plan_datasets}
\title{plan_datasets}
\usage{
plan_datasets(data_path = get_default_data_path())
}
\arguments{
\item{data_path}{path to location of downloaded retriever datasets}
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands 
  for gathering datasets
}
\description{
plan_datasets
}
