% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_from_ldats.R
\name{plot_lda}
\alias{plot_lda}
\title{Plot LDA topic composition and time series}
\usage{
plot_lda(x, observed_dates = NULL, ..., select_samples = NULL)
}
\arguments{
\item{x}{an LDATS LDA object.}

\item{observed_dates}{a vector of dates for the observed samples. If 
observed_dates == NULL, samples will be labelled sequentially starting
from 1.}

\item{...}{additional arguments to pass down}

\item{select_samples}{select samples}
}
\description{
Plot the species composition of LDA topics, and the 
topic composition of the community over the time series.
}
