% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biotime_data.R
\name{prepare_biotime_data}
\alias{prepare_biotime_data}
\title{Prepare Biotime time-series datases for individual loading}
\usage{
prepare_biotime_data(path = get_default_data_path(),
  data_subset = NULL)
}
\arguments{
\item{path}{where to load the raw data files from}

\item{data_subset}{optional, a subset of the Biotime study_ids to use
(to speed up development). As c(1:X)}
}
\description{
Original data found here http://biotime.st-andrews.ac.uk/home.php
}
\examples{
\dontrun{
  get_biotime_data(dataset = 321)
}
}
