% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-retriever-data.R
\name{prepare_datasets}
\alias{prepare_datasets}
\title{Prepare datasets for usage}
\usage{
prepare_datasets(
  dataset = c("breed-bird-survey", "biotimesql"),
  bbs_params = list(path = get_default_data_path(), start_yr = 1965, end_yr = 2018,
    min_num_yrs = 10, data_subset = NULL),
  biotime_params = list(path = get_default_data_path(), data_subset = NULL)
)
}
\arguments{
\item{dataset}{what datasets to prepare (must follow the naming convention
of \code{\link{download_datasets}})}

\item{bbs_params}{params to pass to}

\item{biotime_params}{}
}
\description{
This wraps all the functions that prepare datasets from specific
databases (e.g. \code{\link{prepare_bbs_ts_data}}, \code{\link{prepare_biotime_data}}).
}
