% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_gpdd.R
\name{prepare_gpdd_data}
\alias{prepare_gpdd_data}
\title{Prepare GPDD population time-series data}
\usage{
prepare_gpdd_data(
  path = get_default_data_path(),
  data_subset = NULL,
  min_num_yrs = 10
)
}
\arguments{
\item{path}{where to load the raw data files from}

\item{data_subset}{optional, a subset of the GPDD communities to use
(to speed up development). As c(1:X)}

\item{min_num_yrs}{minimum number of years of data}
}
\description{
Selects sites containing at least \code{min_num_yrs} of data
samples during that period. Cleans data tables and stores each individual
community dataset as a .RDS file. Saves a data table of the \code{location_id}s.
}
