% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matssdata.R
\name{print.matssdata}
\alias{print.matssdata}
\title{Print a time series dataset}
\usage{
\method{print}{matssdata}(x, ...)
}
\arguments{
\item{x}{Class \code{matssdata} object to be printed}

\item{...}{additional arguments (unused)}
}
\description{
Print a time series dataset
}
