% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matssdata.R
\name{print.matsssummary}
\alias{print.matsssummary}
\title{Print a time series summary}
\usage{
\method{print}{matsssummary}(x, ..., n = NULL)
}
\arguments{
\item{x}{Class \code{matsssummary} object to be printed}

\item{...}{additional arguments (unused)}
}
\description{
Print a time series summary
}
