% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biotime_data.R
\name{process_biotime_data}
\alias{process_biotime_data}
\title{Process an individual BioTime dataset}
\usage{
process_biotime_data(biotime_data_tables, dataset_id = 10)
}
\arguments{
\item{biotime_data_tables}{full BioTime data tables}

\item{dataset_id}{the study_id}
}
\value{
the processed BioTime dataset
}
\description{
Filter and modify the BioTime data. Generate the abundance,
covariate, and metadata tables and return the combined object.
}
