% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biotime_data.R
\name{process_biotime_dataset}
\alias{process_biotime_dataset}
\title{Process an individual BioTime dataset}
\usage{
process_biotime_dataset(biotime_data_tables, dataset_id = 10,
  save_to_file = FALSE,
  storage_path = file.path(get_default_data_path(), "biotime-prepped"),
  citation_text = NULL)
}
\arguments{
\item{biotime_data_tables}{full BioTime data tables}

\item{dataset_id}{the study_id}

\item{save_to_file}{whether to save the processed dataset to a file}

\item{storage_path}{folder in which to put processed dataset}

\item{citation_text}{text of citation for the database}
}
\value{
the processed BioTime dataset
}
\description{
Filter and modify the BioTime data. Generate the abundance,
covariate, and metadata tables and return the combined object.
}
