% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_gpdd.R
\name{process_gpdd_data}
\alias{process_gpdd_data}
\title{Process the GPDD data for an individual location and timeperio}
\usage{
process_gpdd_data(
  gpdd_data,
  global_species_table,
  datasource_table,
  save_to_file = FALSE,
  storage_path = file.path(get_default_data_path(), "gpdd-prepped"),
  citation_text = NULL
)
}
\arguments{
\item{global_species_table}{table of species for GPDD}

\item{datasource_table}{table of data sources for GPDD}

\item{save_to_file}{whether to save the processed dataset to a file}

\item{storage_path}{folder in which to put processed dataset}

\item{citation_text}{text of citation for the database}

\item{gpdd}{main GPDD data table}
}
\value{
the processed GPDD dataset
}
\description{
Generate the abundance, covariate, and metadata tables and
return the combined object.
}
