% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_function_sgs.R
\name{process_sgs_data}
\alias{process_sgs_data}
\title{get Shortgrass Steppe rodent data

Import Shortgrass Steppe rodent abundance from data files}
\usage{
process_sgs_data(data_path = here::here("data",
  "shortgrass_steppe_rodents.csv"))
}
\arguments{
\item{data_path}{location of the raw data}
}
\value{
list of two dataframes (one with abundance data, the other with covariate data) 
  and one list of metadata.
}
\description{
get Shortgrass Steppe rodent data

Import Shortgrass Steppe rodent abundance from data files
}
