% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retriever_data.R
\name{retriever_data}
\alias{retriever_data}
\title{Load all retriever data}
\usage{
retriever_data(names = c("bbs", "mapped-plant-quads-mt",
  "fray-jorge-ecology", "veg-plots-sdl"))
}
\arguments{
\item{names}{List of dataset names (from retriever) to load}
}
\value{
A list of lists, all datasets compiled and retaining their retriever names
}
\description{
Load all useable data from retriever
}
\examples{
retriever_data(c("portal","veg-plots-sdl"))

}
