% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{richness}
\alias{richness}
\title{Count non-0 entries}
\usage{
richness(x)
}
\arguments{
\item{x}{\code{numeric} vector}
}
\value{
\code{numeric} value of the number of non-0 entries in \code{x}.
}
\description{
Calculate the richness (number of non-0 entries) of a given 
  sample.
}
