% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda-function.R
\name{run_LDA}
\alias{run_LDA}
\title{run Latent Dirichlet Allocation}
\usage{
run_LDA(data, max_topics = 6, nseeds = 200, ncores = 4)
}
\arguments{
\item{data}{input data, should be a cross-tab of abundances, where each 
row is an observation and each column is a different variable}

\item{max_topics}{maximum number of topics to try in the LDA algorithm}

\item{nseeds}{number of different seeds to try for the method}

\item{ncores}{number of cores to use}
}
\description{
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod 
tempor incididunt ut labore et dolore magna aliqua.
}
