% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_cleaning_functions.R
\name{subset_bbs_route_region_data}
\alias{subset_bbs_route_region_data}
\title{Subset BBS data by route and reigon
Writes each route & region data object (a list of abundance, metadata, and covariates) as an .Rds file to be re-read via readRDS.}
\usage{
subset_bbs_route_region_data(route_region, bbs_data_table, species_table,
  path = get_default_data_path())
}
\arguments{
\item{route_region}{named list of route and region to subset to}

\item{bbs_data_table}{main bbs data table}

\item{species_table}{table of species for BBS}

\item{path}{path}
}
\value{
nothing
}
\description{
Subset BBS data by route and reigon
Writes each route & region data object (a list of abundance, metadata, and covariates) as an .Rds file to be re-read via readRDS.
}
