% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summarize_df}
\alias{summarize_df}
\alias{summarise_df}
\title{Compute summaries and autocorrelation for each variable}
\usage{
summarize_df(
  df,
  times = seq_len(NROW(df)),
  interp_method = forecast::na.interp,
  ...
)

summarise_df(
  df,
  times = seq_len(NROW(df)),
  interp_method = forecast::na.interp,
  ...
)
}
\arguments{
\item{df}{the data.frame of variables to summarize}

\item{times}{the time indices associated with the rows of \code{df}}

\item{interp_method}{\code{character} a function used to interpolate
\code{obs}. Defaults to \code{\link[forecast]{na.interp}}.}

\item{...}{further arguments to be passed to acf}
}
\description{
Compute summaries and autocorrelation for each variable
}
