% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summarize_obs}
\alias{summarize_obs}
\alias{summarise_obs}
\alias{summarize_times}
\alias{summarise_times}
\alias{summarize_effort}
\alias{summarise_effort}
\title{Summarize univariate observations, times, or efforts}
\usage{
summarize_obs(obs, round_out = TRUE, digits = NULL)

summarise_obs(obs, round_out = TRUE, digits = NULL)

summarize_times(obs, times, round_out = TRUE, digits = NULL)

summarise_times(obs, times, round_out = TRUE, digits = NULL)

summarize_effort(obs, effort = rep(1, NROW(obs)), round_out = TRUE,
  digits = NULL)

summarise_effort(obs, effort = rep(1, NROW(obs)), round_out = TRUE,
  digits = NULL)
}
\arguments{
\item{obs}{\code{numeric} vector of observations.}

\item{round_out}{\code{logical} indicator if rounding should happen.}

\item{digits}{\code{NULL} (default) or \code{integer} value of the number
of digits for rounding. If \code{NULL}, \code{digits} is calculated to
be two order of magnitude lower than the smallest value in the vector
being summarized.}

\item{times}{\code{numeric} or \code{Date} vector of timestamps of the
observations.}

\item{effort}{\code{numeric} vector of effort associated with the
observations.}
}
\value{
\code{list} with entries corresponding to the mininum, maximum,
median, mean, standard deviation, and count of the observations, times,
or effort, rounded based on \code{round_out} and \code{digits}.
}
\description{
Summarize univariate observations, times, or efforts
}
