% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summarize_vec}
\alias{summarize_vec}
\title{summarize_vec is a helper function that handles the work of 
  \code{\link{summarize_obs}}, \code{\link{summarize_time}}, and 
  \code{\link{summarize_effort}}}
\usage{
summarize_vec(obs, x, round_out = TRUE, digits = NULL)
}
\arguments{
\item{obs}{\code{numeric} vector of observations.}

\item{x}{the vector to be summarized}

\item{round_out}{\code{logical} indicator if rounding should happen.}

\item{digits}{\code{NULL} (default) or \code{integer} value of the number
of digits for rounding. If \code{NULL}, \code{digits} is calculated to
be two order of magnitude lower than the smallest value in the vector
being summarized.}
}
\description{
summarize_vec is a helper function that handles the work of 
  \code{\link{summarize_obs}}, \code{\link{summarize_time}}, and 
  \code{\link{summarize_effort}}
}
