% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matssdata.R
\name{summary.matssdata}
\alias{summary.matssdata}
\title{Summarize a time series dataset}
\usage{
\method{summary}{matssdata}(x, ...)
}
\arguments{
\item{x}{Class \code{matssdata} object to be summarized.}

\item{...}{additional arguments to \code{ts_summary()}}
}
\value{
\code{list} of number of species, number of
observations, summaries of the variables, the times, the effort, the
species richness, total observation, and the among-species correlation.
}
\description{
Summarize a time series dataset
}
