% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{temp_autocor}
\alias{temp_autocor}
\title{Interpolate observations}
\usage{
temp_autocor(obs, times, interp_method = forecast::na.interp, ...)
}
\arguments{
\item{obs}{a vector, matrix, or data.frame of \code{numeric} observations
(within columns) across times (within rows).}

\item{times}{\code{numeric} or \code{Date} vector of timestamps of the
observations.}

\item{interp_method}{\code{character} a function used to interpolate
\code{obs}. Defaults to \code{\link[forecast]{na.interp}}.}

\item{...}{further arguments to be passed to acf}
}
\value{
Autocorrelation of the observation vector.
}
\description{
Calculate the autocorrelation of the observations based on
their timestamps, with interpolation (based on a specified method) if
necessary.
}
