% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-processing.R
\name{to_numeric_vector}
\alias{to_numeric_vector}
\title{Extract a numeric vector}
\usage{
to_numeric_vector(x)
}
\arguments{
\item{x}{the input data}
}
\description{
Extract a numeric vector from a data.frame or a matrix (taking
the first column).
}
