% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_retriever_data.R
\name{use_default_data_path}
\alias{use_default_data_path}
\title{Manage the default path for downloading MATSS Data into}
\usage{
use_default_data_path(path = NULL)
}
\arguments{
\item{path}{Folder into which data will be downloaded}
}
\description{
See \code{portalr::\link[portalr]{use_default_data_path}} for details.
}
