% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plans.R
\name{build_bbs_datasets_plan}
\alias{build_bbs_datasets_plan}
\title{Generate a Drake Plan for BBS Datasets}
\usage{
build_bbs_datasets_plan(data_path = get_default_data_path(),
  bbs_subset = NULL)
}
\arguments{
\item{data_path}{where to get the downloaded retriever datasets}

\item{bbs_subset}{optional, a subset of the BBS communities to use 
(to speed up development). As c(1:X)}
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands 
  for gathering BBS datasets
}
\description{
Generate a Drake Plan for BBS Datasets
}
