% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plans.R
\name{build_bbs_datasets_plan}
\alias{build_bbs_datasets_plan}
\title{Generate a Drake Plan for BBS Datasets}
\usage{
build_bbs_datasets_plan(data_path = get_default_data_path())
}
\arguments{
\item{path}{path}

\item{from_raw}{whether to re-prep BBS data}
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands 
  for gathering BBS datasets
}
\description{
Generate a Drake Plan for BBS Datasets
}
