% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_data.R
\name{build_biotime_datasets_plan}
\alias{build_biotime_datasets_plan}
\title{Generate a Drake Plan for Biotime Datasets}
\usage{
build_biotime_datasets_plan(
  path = get_default_data_path(),
  data_subset = NULL,
  do_processing = TRUE,
  force_reprocessing = FALSE
)
}
\arguments{
\item{path}{where to get the downloaded retriever datasets}

\item{data_subset}{optional, a subset of the Biotime study_ids to use
(to speed up development). As c(1:X)}

\item{do_processing}{whether to process the datasets if necessary}

\item{force_reprocessing}{whether to force re-processing of datasets}
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands
for gathering Biotime datasets
}
\description{
Generate a Drake Plan for Biotime Datasets
}
