% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plans.R
\name{build_datasets_plan}
\alias{build_datasets_plan}
\title{Generate a Drake Plan for Datasets}
\usage{
build_datasets_plan(data_path = get_default_data_path(),
  include_downloaded_data = FALSE)
}
\arguments{
\item{data_path}{where to get the downloaded retriever datasets}

\item{include_downloaded_data}{whether to also include downloadable datasets}
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands 
  for gathering datasets
}
\description{
Generate a Drake Plan for Datasets
}
