% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plans.R
\name{build_datasets_plan}
\alias{build_datasets_plan}
\title{Generate a Drake Plan for Datasets}
\usage{
build_datasets_plan(data_path = get_default_data_path(),
  include_retriever_data = FALSE, include_bbs_data = FALSE,
  bbs_subset = NULL, include_gpdd_data = FALSE)
}
\arguments{
\item{data_path}{where to get the downloaded retriever datasets}

\item{include_retriever_data}{whether to include retriever-downloaded data}

\item{include_bbs_data}{whether to include BBS data}

\item{bbs_subset}{optional, a subset of the BBS communities to use
(to speed up development). As c(1:X)}

\item{include_gpdd_data}{whether to include gpdd data}
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands
for gathering datasets
}
\description{
Generate a Drake Plan for Datasets
}
