% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plans.R
\name{build_gpdd_datasets_plan}
\alias{build_gpdd_datasets_plan}
\title{Generate a Drake Plan for GPDD Datasets}
\usage{
build_gpdd_datasets_plan(data_path = get_default_data_path())
}
\arguments{
\item{data_path}{where to get the downloaded retriever datasets}
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands
for gathering GPDD datasets
}
\description{
Generate a Drake Plan for GPDD Datasets
}
