% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_analyses.R, R/plan_data.R
\name{build_gpdd_datasets_plan}
\alias{build_gpdd_datasets_plan}
\title{Generate a Drake Plan for GPDD Datasets}
\usage{
build_gpdd_datasets_plan()

build_gpdd_datasets_plan()
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands
for gathering GPDD datasets

a drake plan (i.e. a tibble) specifying the targets and commands
for gathering GPDD datasets
}
\description{
Generate a Drake Plan for GPDD Datasets

Generate a Drake Plan for GPDD Datasets
}
