% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-checks.R
\name{check_data_format}
\alias{check_data_format}
\title{Check data format}
\usage{
check_data_format(data)
}
\arguments{
\item{data}{dataset to check}
}
\value{
TRUE or FALSE
}
\description{
Check whether its input matches the specified data format in
the \code{data-formats.Rmd} vignette
}
