% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_analyses.R
\name{collect_analyses}
\alias{collect_analyses}
\title{Collect the Analyses Together as a List}
\usage{
collect_analyses(list_of_results)
}
\arguments{
\item{list_of_results}{the list of objects}
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands
for all the analyses and the collected results (grouping the outputs from
each method into a single list)
}
\description{
This is a helper function that enables the NSE evaluation in
\code{\link{plan_analyses}}. For example, \code{collect_analyses(list(a, b))}
will return a list of two named elements, \code{a} and \code{b} corresponding to the
objects \code{a} and \code{b}.
}
