% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_cleaning_functions.R
\name{combine_bbs_subspecies}
\alias{combine_bbs_subspecies}
\title{Combine subspecies into their common species}
\usage{
combine_bbs_subspecies(df, species_table)
}
\description{
Modified from https://github.com/weecology/bbs-forecasting/blob/master/R/forecast-bbs-core.R
}
