% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_retriever_data.R
\name{get_bbs_data}
\alias{get_bbs_data}
\title{Create BBS population time-series data}
\usage{
get_bbs_data(start_yr = 1965, end_yr = 2017, min_num_yrs = 10,
  region, path = get_default_data_path())
}
\arguments{
\item{start_yr}{num first year of time-series}

\item{end_yr}{num last year of time-series}

\item{min_num_yrs}{num minimum number of years of data between start_yr & end_yr}

\item{region}{region code of data to return (currently uses state codes)}

\item{path}{where to load the raw data files from}
}
\value{
list of two dataframes (one with abundance data, the other with covariate data) 
  and one list of metadata.
}
\description{
Modified from https://github.com/weecology/bbs-forecasting
and https://github.com/weecology/MATSS-community-change 
Selects sites with data spanning start_yr through end_yr containing at least min_num_yrs of data
samples during that period.
}
\examples{
\dontrun{
  get_bbs_data(region=7)
}
}
