% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_retriever_data.R
\name{get_bbs_route_region_data}
\alias{get_bbs_route_region_data}
\title{Get BBS data by route and reigon}
\usage{
get_bbs_route_region_data(route, region, bbs_ts_data)
}
\arguments{
\item{route}{route}

\item{region}{region}

\item{bbs_ts_data}{list of three dataframes (output of prepare_bbs_ts_data)}
}
\value{
list of two dataframes (one with abundance data, the other with covariate data) 
  and one list of metadata.
}
\description{
Get BBS data by route and reigon
}
