% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_gpdd.R
\name{get_gpdd_data}
\alias{get_gpdd_data}
\title{Get a cleaned GPDD dataset}
\usage{
get_gpdd_data(
  path = file.path(get_default_data_path(), "gpdd-prepped", paste0("location",
    location_id, ".RDS")),
  location_id = 83
)
}
\arguments{
\item{path}{where to load the raw data files from}

\item{location_id}{Location code of data to return}

\item{timeperiod_id}{Sampling timescale code of data to return
(some datasets provide at more than one scale)}
}
\value{
list of abundance, covariates, and metadata
}
\description{
Gets a prepped GPDD dataset (as a list of abundance,
covariates, and metadata) for a specified \code{location_id} and \code{timeperiod_id}
First run \code{\link{prepare_gpdd_data}} to create these files from the raw
GPDD database. If the files are not found, then \code{NULL} is returned.
}
