% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_gpdd.R
\name{get_gpdd_data}
\alias{get_gpdd_data}
\title{Create GPDD population time-series data}
\usage{
get_gpdd_data(
  path = get_default_data_path(),
  location_id = 83,
  timeperiod_id = 408,
  min_num_yrs = 10
)
}
\arguments{
\item{path}{where to load the raw data files from}

\item{location_id}{Location code of data to return}

\item{timeperiod_id}{Sampling timescale code of data to return
(some datasets provide at more than one scale)}

\item{min_num_yrs}{minimum number of years of data}
}
\value{
list of abundance, covariates, and metadata
}
\description{
Selects sites containing at least `min_num_yrs`` of data
samples during that period.
}
\examples{
\dontrun{
  get_gpdd_data(location_id=83, timeperiod_id=408)
}
}
