% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_jornada.R
\name{get_jornada_data}
\alias{get_jornada_data}
\title{get Jornada rodent data

Import Jornada rodent abundance from data files}
\usage{
get_jornada_data(
  path = file.path(get_default_data_path(), "jornada-lter-rodent")
)
}
\arguments{
\item{path}{where to load the raw data files from}
}
\value{
list of abundance, covariates, and metadata
}
\description{
get Jornada rodent data

Import Jornada rodent abundance from data files
}
