% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_maizuru.R
\name{get_maizuru_data}
\alias{get_maizuru_data}
\title{get the maizuru community data}
\usage{
get_maizuru_data(
  path = file.path(get_default_data_path(), "ushio-maizuru-fish-community")
)
}
\arguments{
\item{path}{where to load the raw data files from}
}
\value{
list of abundance, covariates, and metadata
}
\description{
get the maizuru community data
}
