% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maizuru_data.R
\name{get_maizuru_data}
\alias{get_maizuru_data}
\title{Read in the maizuru community data from retriever downloaded CSV

Import maizuru data from data files}
\usage{
get_maizuru_data(path = file.path(get_default_data_path(),
  "ushio-maizuru-fish-community"))
}
\value{
list of two dataframes (one with abundance data, the other with
covariate data) and one list of metadata.
}
\description{
Read in the maizuru community data from retriever downloaded CSV

Import maizuru data from data files
}
