% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_retriever.R
\name{get_mtquad_data}
\alias{get_mtquad_data}
\title{Create Montana plant quad time-series data}
\usage{
get_mtquad_data(path = file.path(get_default_data_path(),
  "mapped-plant-quads-mt"))
}
\arguments{
\item{path}{where to load the raw data files from}
}
\value{
list of two dataframes (one with abundance data, the other with covariate data)
and one list of metadata.
}
\description{
Original data found here
}
\examples{
\dontrun{
  get_mtquad_data(mtquad_data_tables=retriever_data()$'mapped-plant-quads-mt')
}
}
