% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{get_mtquad_data}
\alias{get_mtquad_data}
\title{Create Montana plant quad time-series data}
\usage{
get_mtquad_data(mtquad_data_tables)
}
\arguments{
\item{mtquad_data_tables}{list of all montana plant tables (allrecords_cover,
allrecords_density,species_list)}
}
\value{
list of two dataframes, one with abundance data, the other with covariate data
}
\description{
Original data found here
}
\examples{
get_mtquad_data(mtquad_data_tables=retriever_data()$'mapped-plant-quads-mt')
}
