% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_vegplots-sdl.R
\name{get_sdl_data}
\alias{get_sdl_data}
\title{Create Sonoran desert lab time-series data}
\usage{
get_sdl_data(
  plots = c(4, 7, 8, 9, 10, 11, 12, 14, 15, 16, 17),
  path = file.path(get_default_data_path(), "veg-plots-sdl")
)
}
\arguments{
\item{plots}{vector of plots to keep}

\item{path}{where to load the raw data files from}
}
\value{
list of abundance, covariates, and metadata
}
\description{
Original data found here http://www.eebweb.arizona.edu/faculty/venable/LTREB/LTREB\%20data.htm
}
