% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_retriever_data.R
\name{get_sdl_data}
\alias{get_sdl_data}
\title{Create Sonoran desert lab time-series data}
\usage{
get_sdl_data(plots = c(4, 7, 8, 9, 10, 11, 12, 14, 15, 16, 17),
  path = get_default_data_path())
}
\arguments{
\item{plots}{vector of plots to keep}

\item{path}{where to load the raw data files from}
}
\value{
list of two dataframes (one with abundance data, the other with covariate data) 
  and one list of metadata.
}
\description{
Original data found here http://www.eebweb.arizona.edu/faculty/venable/LTREB/LTREB%20data.htm
}
\examples{
\dontrun{
  get_sdl_data(sdl_data_tables=retriever_data()$'veg-plots-sdl')
}
}
