% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sgs_data.R
\name{get_sgs_data}
\alias{get_sgs_data}
\title{get Shortgrass Steppe rodent data

Import Shortgrass Steppe rodent abundance from data files}
\usage{
get_sgs_data()
}
\value{
list of two dataframes (one with abundance data, the other with 
  covariate data) and one list of metadata.
}
\description{
get Shortgrass Steppe rodent data

Import Shortgrass Steppe rodent abundance from data files
}
