% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-retriever-data.R
\name{install_retriever_data}
\alias{install_retriever_data}
\alias{import_retriever_data}
\alias{download_datasets}
\title{Download data from the data retriever}
\usage{
install_retriever_data(dataset, path = get_default_data_path(),
  force_install = FALSE)

import_retriever_data(dataset = NULL, path = get_default_data_path())

download_datasets(dataset = c("breed-bird-survey", "veg-plots-sdl",
  "mapped-plant-quads-mt", "biotimesql"), path = get_default_data_path(),
  force_install = FALSE)
}
\arguments{
\item{dataset}{the name of the dataset that you wish to download}

\item{path}{the overarching folder in which to download datasets}

\item{force_install}{whether to install the dataset if the correctly named
folder already exists}

\item{path}{the overarching folder in which to download datasets; OR the
full path to the folder containing the data (when \code{dataset == NULL})}
}
\description{
\code{install_retriever_data} downloads retriever datasets and
is a wrapper around rdataretriever::install

\code{import_retriever_data} loads a previously downloaded
retriever dataset

\code{download_datasets} is a wrapper around
\code{\link{install_retriever_data}} to download multiple datasets, with
the default to download all of the datasets that are supported.
}
\examples{
\dontrun{
  install_retriever_data("veg-plots-sdl")
}
\dontrun{
  import_retriever_data("veg-plots-sdl")
}
\dontrun{
  download_datasets()
}
}
