% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_retriever_data.R
\name{install_retriever_data}
\alias{install_retriever_data}
\title{Download a dataset from data retriever}
\usage{
install_retriever_data(dataset, data_path = "data")
}
\arguments{
\item{dataset}{the name of the dataset that you wish to download}

\item{data_path}{the overarching folder in which to download datasets}
}
\description{
a wrapper around rdataretriever::install to facilitate 
  downloading multiple datasets into a common data folder
}
\examples{
\dontrun{
  install_retriever_data("veg-plots-sdl")
}
}
