% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{interpolate_obs}
\alias{interpolate_obs}
\title{Interpolate observations}
\usage{
interpolate_obs(obs, times, method = "na.interp")
}
\arguments{
\item{obs}{\code{numeric} vector of observations.}

\item{times}{\code{numeric} or \code{Date} vector of timestamps of the 
observations.}

\item{method}{\code{character} representing a function name used to 
interpolate \code{obs}. Defaults to \code{"\link[forecast]{na.interp}"}.}
}
\value{
Interpolated observation vector.
}
\description{
Interpolate observations based on their timestamps and a 
  method.
}
