% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{interpolate_obs}
\alias{interpolate_obs}
\title{Interpolate observations}
\usage{
interpolate_obs(obs, times, method = "na.interp", ...)
}
\arguments{
\item{obs}{\code{ts_summary}: a 1- or 2-dimensional set of \code{numeric}
observations (columns if 2-dimensional) across times (rows if 
2-dimensional) as indexed by \code{times}. \cr \cr
\code{uni_ts_summary}: a 1-dimensional \code{numeric} set of observations
across times as indexed by \code{times}.}

\item{times}{\code{numeric} or \code{Date} vector of timestamps of the 
observations.}

\item{...}{further arguments to be passed to the interpolation method}
}
\value{
Interpolated observation vector.
}
\description{
Interpolate observations based on their timestamps and a 
  method.
}
