% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maizuru_data.R
\name{maizuru_data}
\alias{maizuru_data}
\title{Dataset Maizuru Fish Community}
\description{
A dataset containing the twice-monthly visual census data of the Maizuru fish 
  community (only the 15 species that had total observation count > 1000).
Data are available from https://zenodo.org/record/1181937
}
