% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATSS-package.R
\docType{package}
\name{MATSS}
\alias{MATSS}
\alias{MATSS-package}
\title{Macroecological Analayses of Time Series Structure}
\description{
Functions for obtaining and processing ecological time series 
  datasets, and for running different analyses.
}
\keyword{package}
