% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-summary-stats-conversions.R
\name{normalize_obs}
\alias{normalize_obs}
\alias{normalize_effort,}
\alias{normalize_times}
\alias{normalize_effort}
\title{Normalize observations, effort, or times}
\usage{
normalize_obs(obs, effort, obs_per_effort = !is.null(effort))

normalize_times(obs, times = NULL)

normalize_effort(obs, effort = NULL)
}
\arguments{
\item{obs}{the time series of \code{numeric} observations}

\item{effort}{\code{numeric} vector of effort associated with the
observations.}

\item{obs_per_effort}{\code{logical} indicator if \code{obs} should be
corrected for \code{effort} before summaries are done.}

\item{times}{\code{numeric} or \code{Date} vector of timestamps of the
observations.}
}
\description{
\code{normalize_obs} will normalize the observations to the
provided effort vector, if \code{obs_per_effort} is TRUE.

\code{normalize_effort} will generate a default effort vector if it is
not provided, (effort = 1 over the provided \code{obs} time series)

\code{normalize_times} will generate a default times vector if it is
not provided, (times = \code{seq(length(obs))})
}
