% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_bbs.R
\name{prepare_bbs_ts_data}
\alias{prepare_bbs_ts_data}
\title{Prepare BBS population time-series data}
\usage{
prepare_bbs_ts_data(
  start_yr = 1965,
  end_yr = 2018,
  min_num_yrs = 10,
  path = get_default_data_path(),
  data_subset = NULL
)
}
\arguments{
\item{start_yr}{num first year of time-series}

\item{end_yr}{num last year of time-series}

\item{min_num_yrs}{num minimum number of years of data between start_yr & end_yr}

\item{path}{where to load the raw data files from}

\item{data_subset}{optional, a subset of the BBS communities to use
(to speed up development). As c(1:X)}
}
\description{
Modified from \url{https://github.com/weecology/bbs-forecasting}
and \url{https://github.com/weecology/MATSS-community-change}.

Selects sites with data spanning \code{start_yr} through \code{end_yr} containing at
least \code{min_num_yrs} of data samples during that period. Cleans data tables
and stores each individual route as a .RDS file. Saves a data table of the
route + region pairs.
}
