% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_cleaning_functions.R
\name{process_bbs_route_region_data}
\alias{process_bbs_route_region_data}
\title{Process the BBS data for an individual route and region}
\usage{
process_bbs_route_region_data(bbs_data_table, species_table)
}
\arguments{
\item{bbs_data_table}{main bbs data table}

\item{species_table}{table of species for BBS}
}
\value{
the processed BBS data
}
\description{
Correct and otherwise filter BBS species data (see
\code{\link{combine_bbs_subspecies}} and \code{\link{filter_bbs_species}}
for more info). Generate the abundance, covariate, and metadata tables and
return the combined object.
}
