% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_jornada_rodents.R
\name{process_jornada_data}
\alias{process_jornada_data}
\title{get Jornada rodent data

Import Jornada rodent abundance from data files}
\usage{
process_jornada_data()
}
\description{
get Jornada rodent data

Import Jornada rodent abundance from data files
}
