% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_function_sgs.R
\name{process_sgs_data}
\alias{process_sgs_data}
\title{get Shortgrass Steppe rodent data

Import Shortgrass Steppe rodent abundance from data files}
\usage{
process_sgs_data()
}
\arguments{
\item{data_path}{location of the raw data}
}
\description{
get Shortgrass Steppe rodent data

Import Shortgrass Steppe rodent abundance from data files
}
