% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summarize_vec}
\alias{summarize_vec}
\alias{summarise_vec}
\title{Summarize a univariate vector}
\usage{
summarize_vec(x, round_out = TRUE, digits = NULL)

summarise_vec(x, round_out = TRUE, digits = NULL)
}
\arguments{
\item{x}{the vector to be summarized}

\item{round_out}{\code{logical} indicator if rounding should happen.}

\item{digits}{\code{NULL} (default) or \code{integer} value of the number
of digits for rounding. If \code{NULL}, \code{digits} is calculated to
be two order of magnitude lower than the smallest value in the vector
being summarized.}
}
\value{
\code{vector} with entries corresponding to the mininum, maximum,
median, mean, standard deviation, and count of the observations, times,
or effort, rounded based on \code{round_out} and \code{digits}.
}
\description{
Summarize a univariate vector
}
