% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matssdata.R
\name{summary.matssdata}
\alias{summary.matssdata}
\title{Summarize a time series dataset}
\usage{
\method{summary}{matssdata}(
  data,
  times = NULL,
  effort = NULL,
  obs_per_effort = !is.null(effort),
  interp_method = forecast::na.interp,
  include_spp_correlations = TRUE,
  ...
)
}
\arguments{
\item{...}{additional arguments to \code{ts_summary()}}

\item{x}{Class \code{matssdata} object to be summarized.}
}
\value{
\code{list} of number of species, number of
observations, summaries of the variables, the times, the effort, the
species richness, total observation, and the among-species correlation.
}
\description{
Summarize a time series dataset
}
