% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{temp_autocor}
\alias{temp_autocor}
\title{Interpolate observations}
\usage{
temp_autocor(obs, times, interp_method = "na.interp")
}
\arguments{
\item{obs}{\code{numeric} vector of observations.}

\item{times}{\code{numeric} or \code{Date} vector of timestamps of the 
observations.}

\item{interp_method}{\code{character} representing a function name used to 
interpolate \code{obs}. Defaults to \code{"\link[forecast]{na.interp}"}.}
}
\value{
Autocorrelation of the observation vector.
}
\description{
Calculate the autocorrelation of the observations based on 
  their timestamps, with interpolation (based on a specified method) if 
  necessary.
}
